+++ springboot/src/main/java/com/TreeNut/ChatBot_Backend/service/CharacterService.kt
